[삭제] controller에서 다오 사용부분 제거, verifyToken을 auth.js로 합치기
