+++ middleWare/auth.js
@@ -1,4 +1,7 @@
 import verifyToken from './token';
+import jwt from 'jsonwebtoken';
+import dotenv from 'dotenv';
+dotenv.config();
 
 //인증
 const authentication = (req, res, next) => {
@@ -15,4 +18,13 @@ const authentication = (req, res, next) => {
   }
 };
 
-export default authentication;
+// 토큰 확인 절차
+const verifyToken = token => {
+  try {
+    return jwt.verify(token, process.env.JWT_SECRET);
+  } catch (err) {
+    return null;
+  }
+};
+
+export default { authentication, verifyToken };
